requirements.txt
diffusers
kombu
torchdynamo
pillow